Refactor store into a lib folder
